put search page toolbar in the search screen
