+++ app/src/main/java/com/github/tinkzhang/readkeeper/common/DataStore.kt
@@ -1,0 +1,8 @@
+package com.github.tinkzhang.readkeeper.common
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStore
+
+val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "settings")
