+++ app/src/main/java/com/github/tinkzhang/readkeeper/search/SearchResultViewModel.kt
@@ -21,6 +21,7 @@ class SearchResultViewModel(private val keyword: String) : ViewModel() {
 
     fun addHistory(keyword: String) {
         // TODO: add search history
+
     }
 }
