+++ app/src/main/java/com/github/tinkzhang/readkeeper/search/SearchViewModel.kt
@@ -1,0 +1,6 @@
+package com.github.tinkzhang.readkeeper.search
+
+import androidx.lifecycle.ViewModel
+
+class SearchViewModel : ViewModel(){
+}
